requirements.txt
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.0
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
Django==6.0
djangorestframework==3.16.1
gunicorn==25.3.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jiter==0.12.0
joblib==1.5.3
nltk==3.9.3
openai==2.30.0
packaging==26.0
pillow==12.1.1
psycopg2==2.9.11
pycryptodome==3.23.0
pydantic==2.12.5
pydantic_core==2.41.5
regex==2026.2.19
requests==2.32.5
sniffio==1.3.1
sqlparse==0.5.4
tqdm==4.67.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.6.3
